Initial version of get_unread_books
As yet untested, but let's get a checkpoint.

Given an amazon username, amazon password, and an unread percentage, uses
the 'lector' Kindle API to grab the list of books in your library, find
the ones where you've read less than or equal to the unread percentage,
and print that list.